requirements.txt
streamlit==1.38.0
langchain==0.2.14
langchain-community==0.2.12
faiss-cpu==1.12.0
pymupdf==1.26.4
pandas==2.2.2
numpy==1.26.4
openpyxl
xlrd
langchain-google-genai
python-dotenv
